Moving builder around and seeing how it feels
