--- src/builder/components.go
@@ -1,1 +1,0 @@
-package builder
