+++ src/display/application.go
@@ -1,5 +1,7 @@
 package display
 
+/*
+
 import (
 	"github.com/go-gl/gl/v2.1/gl"
 	"github.com/go-gl/glfw/v3.2/glfw"
@@ -187,3 +189,4 @@ func ApplicationLoop(d Displayable) {
 		time.Sleep(waitDuration)
 	}
 }
+*/
