+++ src/display/builder.go
@@ -1,161 +1,93 @@
 package display
 
-// import "errors"
-
-type SurfaceType int
-
-const (
-	CairoSurfaceType = iota
-	ImageSurfaceType
-	FakeSurfaceType
+import (
+	"errors"
 )
-
-type GlfwWindowHint int
-
-const (
-	AutoIconify = iota
-	Decorated
-	Floating
-	Focused
-	Iconified
-	Maximized
-	Resizable
-	Visible
-)
-
-/*
-// TODO(lbayes) Map local hints to glfw library hints
-const (
-	Focused     GlfwWindowHint = C.GLFW_FOCUSED      // Specifies whether the window will be given input focus when created. This hint is ignored for full screen and initially hidden windows.
-	Iconified   Hint = C.GLFW_ICONIFIED    // Specifies whether the window will be minimized.
-	Maximized   Hint = C.GLFW_MAXIMIZED    // Specifies whether the window is maximized.
-	Visible     Hint = C.GLFW_VISIBLE      // Specifies whether the window will be initially visible.
-	Resizable   Hint = C.GLFW_RESIZABLE    // Specifies whether the window will be resizable by the user.
-	Decorated   Hint = C.GLFW_DECORATED    // Specifies whether the window will have window decorations such as a border, a close widget, etc.
-	Floating    Hint = C.GLFW_FLOATING     // Specifies whether the window will be always-on-top.
-	AutoIconify Hint = C.GLFW_AUTO_ICONIFY // Specifies whether fullscreen windows automatically iconify (and restore the previous video mode) on focus loss.
-)
-
-*/
-
-type Builder2 interface {
-	// Provide a named surface type (CairoSurface, ImageSurface, FakeSurface)
-	WithSurfaceType(t SurfaceType) Builder2
-	// Provide a concrete surface instance.
-	WithSurface(t Surface) Builder2
-	// Provide the desired frame rate (in frames per second)
-	WithFrameRate(fps int) Builder2
-	// Provide zero or more Window Hints to Glfw.
-	WithHint(name GlfwWindowHint, value interface{}) Builder2
-	// Provide width and height to the native window.
-	WithSize(width int, height int) Builder2
-	// Provide title to the native window.
-	WithTitle(title string) Builder2
-	// Execute the configured builder and receive a surface that can be used to
-	//draw visual components.
-	Build(handler func(s Surface)) error
-
-	GetNativeWindowSize() (width int, height int)
-	PollEvents() []interface{}
-}
-
-type builder2 struct {
-	windowHints map[GlfwWindowHint]interface{}
-}
-
-// Provide a named surface type (CairoSurface, ImageSurface, FakeSurface)
-func (b *builder2) WithSurfaceType(t SurfaceType) Builder2 {
-	return b
-}
-
-// Provide a concrete surface instance.
-func (b *builder2) WithSurface(t Surface) Builder2 {
-	return b
-}
-
-// Provide the desired frame rate (in frames per second)
-func (b *builder2) WithFrameRate(fps int) Builder2 {
-	return b
-}
-
-// Provide zero or more Window Hints to Glfw.
-func (b *builder2) WithHint(name GlfwWindowHint, value interface{}) Builder2 {
-	return b
-}
-
-// Provide width and height to the native window.
-func (b *builder2) WithSize(width int, height int) Builder2 {
-	return b
-}
-
-// Provide title to the native window.
-func (b *builder2) WithTitle(title string) Builder2 {
-	return b
-}
-
-// Execute the configured builder and receive a surface that can be used to
-//draw visual components.
-func (b *builder2) Build(handler func(s Surface)) error {
-	return nil
-}
-
-func (b *builder2) GetNativeWindowSize() (width int, height int) {
-	return 0, 0
-}
-
-func (b *builder2) PollEvents() []interface{} {
-	return nil
-}
-
-func NewBuilder2() Builder2 {
-	return &builder2{}
-}
-
-//----------------------------------------------------------------------
 
 type Builder interface {
-	Build()
-	BuildWithRoot(d Displayable)
-	GetRoot() Displayable
+	Build(factory ComponentComposer) (root Displayable, err error)
+	GetFrameRate() int
+	GetHeight() int
+	GetSize() (width int, height int)
+	GetSurfaceType() SurfaceTypeName
+	GetTitle() string
+	GetWidth() int
+	GetWindowHint(hintName GlfwWindowHint) interface{}
+	GetWindowHints() []*windowHint
+	Push(d Displayable)
 }
 
-// Factory that operates over semantic sugar that we use to describe the
-// displayable hierarchy.
 type builder struct {
-	*SurfaceDelegate
-	stack        Stack
-	root         Displayable
-	buildHandler func(Surface)
+	surfaceTypeName SurfaceTypeName
+	frameRate       int
+	windowHints     []*windowHint
+	width           int
+	height          int
+	title           string
+	root            Displayable
+	stack           Stack // TODO: Move THIS stack def into builder package
+	lastError       error
 }
 
-func (b *builder) getStack() Stack {
-	if b.stack == nil {
-		b.stack = NewStack()
+func (b *builder) applyDefaults() {
+	b.frameRate = DefaultFrameRate
+	b.width = DefaultWidth
+	b.height = DefaultHeight
+	b.title = DefaultTitle
+	b.applyDefaultWindowHints()
+	b.stack = NewStack()
+}
+
+func (b *builder) applyDefaultWindowHints() {
+	b.windowHints = []*windowHint{
+		&windowHint{name: AutoIconify, value: false},
+		&windowHint{name: Decorated, value: false},
+		&windowHint{name: Floating, value: true},
+		&windowHint{name: Focused, value: true},
+		&windowHint{name: Iconified, value: false},
+		&windowHint{name: Maximized, value: false},
+		&windowHint{name: Resizable, value: true},
+		&windowHint{name: Visible, value: true},
 	}
-	return b.stack
 }
 
-func (b *builder) GetRoot() Displayable {
-	return b.root
+func (b *builder) removeWindowHint(hintName GlfwWindowHint) {
+	hints := b.windowHints
+	for i := 0; i < len(hints); i++ {
+		if hints[i].name == hintName {
+			b.windowHints = append(hints[:i], hints[i+1:]...)
+			return
+		}
+	}
 }
 
-func (b *builder) Push(d Displayable) error {
+func (b *builder) createSurface() Surface {
+	// create new surface here
+	return nil
+}
+
+func (b *builder) Push(d Displayable) {
 	if b.root == nil {
 		b.root = d
 	}
 
-	s := b.getStack()
-	parent := s.Peek()
+	// Get the parent element if one exists
+	parent := b.stack.Peek()
 
-	if parent != nil {
+	if parent == nil {
+		if b.root != d {
+			// It looks like we have a second root definition in the outer factory function
+			b.lastError = errors.New("Component factory function should only have a single root node")
+			return
+		}
+	} else {
 		parent.AddChild(d)
 	}
 
-	err := s.Push(d)
-	if err != nil {
-		return err
-	}
+	// Push the element onto the stack
+	b.stack.Push(d)
 
+	// Render the element children by calling it's compose function
 	decl := d.GetDeclaration()
 	if decl.Compose != nil {
 		decl.Compose()
@@ -163,33 +95,83 @@ func (b *builder) Push(d Displayable) error {
 		panic("Not yet implemented")
 	}
 
-	//if b.root == d {
-	// d.builder(r)
-	// } else {
-	s.Pop()
-	// }
+	// Pop the element off the stack
+	b.stack.Pop()
+}
 
+func (b *builder) Build(factory ComponentComposer) (root Displayable, err error) {
+	// We may have a configuration error that was stored for later. If so, stop
+	// and return it now.
+	if b.lastError != nil {
+		return nil, err
+	}
+
+	factory(b)
+
+	if b.lastError != nil {
+		return nil, b.lastError
+	}
+
+	return b.root, nil
+}
+
+func (b *builder) GetSurfaceType() SurfaceTypeName {
+	return b.surfaceTypeName
+}
+
+func (b *builder) GetFrameRate() int {
+	return b.frameRate
+}
+
+func (b *builder) GetWindowHint(hintName GlfwWindowHint) interface{} {
+	for _, hint := range b.windowHints {
+		if hint.name == hintName {
+			return hint.value
+		}
+	}
 	return nil
 }
 
-func (b *builder) Reset() {
-	b.stack = NewStack()
-	b.root = nil
+func (b *builder) GetWindowHints() []*windowHint {
+	return b.windowHints
 }
 
-func (b *builder) BuildWithRoot(root Displayable) {
-	b.Reset()
-	b.Push(root)
-	b.buildHandler(b)
-	b.root.Render(b)
+func (b *builder) GetWidth() int {
+	return b.width
 }
 
-func (b *builder) Build() {
-	b.Reset()
-	b.buildHandler(b)
-	b.root.Render(b)
+func (b *builder) GetHeight() int {
+	return b.height
 }
 
-func CreateBuilder(s Surface, buildHandler func(s Surface)) Builder {
-	return &builder{SurfaceDelegate: NewSurfaceDelegate(s), buildHandler: buildHandler}
+func (b *builder) GetSize() (width, height int) {
+	return b.width, b.height
+}
+
+func (b *builder) GetTitle() string {
+	return b.title
+}
+
+// Create a new builder instance with the provided options.
+// This pattern was discovered by Rob Pike and published here:
+// https://commandcenter.blogspot.com/2014/01/self-referential-functions-and-design.html
+// It was also supported by Dave Cheney here:
+// https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis
+// and here:
+// https://dave.cheney.net/2016/11/13/do-not-fear-first-class-functions
+// I'm exploring the idea and finding it to be pretty compelling, especially for what
+// we'd like to consider "immutable" values.
+func NewBuilder(args ...BuilderOption) Builder {
+	b := &builder{}
+	b.applyDefaults()
+
+	for _, arg := range args {
+		err := arg(b)
+		if err != nil {
+			b.lastError = err
+			break
+		}
+	}
+
+	return b
 }
