+++ src/examples/boxes/main.go
@@ -2,16 +2,20 @@ package main
 
 import . "display"
 
-func CreateBoxesApp() Displayable {
-	return Application(&Opts{Title: "Example"}, func(s Surface) {
-		Box(s, func() {
-			// Box(s, FlexWidth(1), FlexHeight(1), MaxWidth(321), MaxHeight(2423))
-			Box(s, &Opts{FlexWidth: 1, FlexHeight: 1, MaxWidth: 640, MaxHeight: 480})
-			Box(s, &Opts{FlexWidth: 1, FlexHeight: 1, MaxWidth: 320, MaxHeight: 280})
-		})
+func CreateBoxesApp() (Displayable, error) {
+	return NewBuilder(Title("Hello World"), Size(640, 480)).Build(func(b Builder) {
+		Sprite(b)
+		// Box(s, FlexWidth(1), FlexHeight(1), MaxWidth(321), MaxHeight(2423))
+		// Box(b, &Opts{FlexWidth: 1, FlexHeight: 1, MaxWidth: 640, MaxHeight: 480})
+		// Box(b, &Opts{FlexWidth: 1, FlexHeight: 1, MaxWidth: 320, MaxHeight: 280})
+		// })
 	})
 }
 
 func main() {
-	ApplicationLoop(CreateBoxesApp())
+	_, err := CreateBoxesApp()
+	if err != nil {
+		panic(err)
+	}
+
 }
