+++ src/examples/boxes/main_test.go
@@ -10,7 +10,7 @@ func TestBoxesMain(t *testing.T) {
 
 	t.Run("", func(t *testing.T) {
 		surface := &display.FakeSurface{}
-		root := CreateBoxesApp()
+		root, _ := CreateBoxesApp()
 		root.Render(surface)
 		// NOTE(lbayes): This should be 1, but is not yet implemented
 		assert.Equal(root.GetChildCount(), 0)
